// web/src/api/generated/model/taxCodeInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Baba.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { TaxTreatment } from './taxTreatment';

export interface TaxCodeInput {
  code: string;
  nameAr: string;
  nameEn: string;
  rate: number;
  treatment: TaxTreatment;
  /** @nullable */
  effectiveFrom: string | null;
  /** @nullable */
  effectiveTo: string | null;
  /** @nullable */
  outputAccountId: string | null;
  /** @nullable */
  inputAccountId: string | null;
}