// web/src/api/generated/model/taxTreatment.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Baba.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export type TaxTreatment = typeof TaxTreatment[keyof typeof TaxTreatment];


export const TaxTreatment = {
  Standard: 'Standard',
  Zero: 'Zero',
  Exempt: 'Exempt',
  OutOfScope: 'OutOfScope',
} as const;